System.MIDI

Description

A lowest common denominator interface to the Win32 and MacOSX MIDI bindings. Error handling is via fail-s in the IO monad. .

Always link with the threaded runtime! (use the -threaded GHC option)

Documentation

module System.MIDI.Base

type Source = Source

The opaque data type representing a MIDI source.

type Destination = Destination

The opaque data type representing a MIDI destination.

type Connection = Connection

The opaque data type representing a MIDI connection.

enumerateSources :: IO [Source]

Enumerates the MIDI sources present in the system.

enumerateDestinations :: IO [Destination]

Enumerates the MIDI destinations present in the system.

class MIDIHasName c

getName :: MIDIHasName a => a -> IO String

These functions return the name, model and manufacturer of a MIDI source / destination.

Note: On Win32, only getName returns a somewhat meaningful string at the moment.

getModel :: MIDIHasName a => a -> IO String

getManufacturer :: MIDIHasName a => a -> IO String

openSource :: Source -> Maybe ClientCallback -> IO Connection

Opens a MIDI Source. There are two possibilites to receive MIDI messages. The user can either supply a callback function, or get the messages from an asynchronous buffer. However, mixing the two approaches is not allowed.

openDestination :: Destination -> IO Connection

Opens a MIDI Destination.

close :: Connection -> IO ()

Closes a MIDI Connection.

send :: Connection -> MidiMessage -> IO ()

Sends a short message. The connection must be a Destination.

sendSysEx :: Connection -> [Word8] -> IO ()

Sends a system exclusive message. You should not include the starting / trailing bytes 0xF0 and 0xF7.

Note: On Win32, the connection must be a Destination

start :: Connection -> IO ()

Starts a connection. This is required for receiving MIDI messages, and also for starting the clock.

stop :: Connection -> IO ()

Stops a connection.

getNextEvent :: Connection -> IO (Maybe MidiEvent)

Gets the next event from a buffered connection (see also openSource)

checkNextEvent :: Connection -> IO (Maybe MidiEvent)

Checks the next event from a buffered connection, but does not remove it from the buffer.

getEvents :: Connection -> IO [MidiEvent]

Gets all the events from the buffer (see also openSource)

getEventsUntil :: Connection -> TimeStamp -> IO [MidiEvent]

Gets all the events with timestamp less than the specified from the buffer.

currentTime :: Connection -> IO Word32

Returns the time elapsed since the last start call, in milisecs.