System.MIDI.Utility

Description

Helper functions to make it easy to start messing around

Documentation

selectMidiDevice

Arguments

:: MIDIHasName a

=> String
prompt

-> Maybe String
default device name

-> [a]
list of devices

-> IO a

Utility function to help choosing a midi device. If there is only a single device, we select that. You can also set a default device (by its name), which will be automatically selected if present.

selectInputDevice :: String -> Maybe String -> IO Source

Select a MIDI input device (source)

selectOutputDevice :: String -> Maybe String -> IO Destination

Select a MIDI output device (destination)